requirements.txt
asgiref==3.8.1
Django==5.0.8
django-browser-reload==1.13.0
django-extensions==3.2.3
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
